% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Generic.R
\name{as.geospatial.coordinate}
\alias{as.geospatial.coordinate}
\title{as.geospatial.coordinate wrapper function}
\usage{
as.geospatial.coordinate(latitude, longitude, elevation, ...)
}
\arguments{
\item{latitude}{the latitude of a geospatial coordinate}

\item{longitude}{of a geospatial coordinate}
}
\description{
This function is a wrapper which formats a latitude, longitude pair as a 
 geospatial coordinate.  This is useful if a coordinate is treated as a single 
 parameter
 
 The parameters can be shorted to lat, lon if space is short.
}
\examples{

as.geospatial.coordinate(lat = 10, lon = 5)

}
\keyword{time}
