% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Generic.R
\name{as.International.Date}
\alias{as.International.Date}
\title{as.International.Date wrapper function}
\usage{
as.International.Date(date_text)
}
\arguments{
\item{date_text}{the text to convert into a date}
}
\description{
This function is a wrapper for the as.Date function.
 
 The necessity of the format string is gone
}
\examples{

as.International.Date("1992/1/1")

}
\keyword{time}
