% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Generic.R
\name{as.Year}
\alias{as.Year}
\title{as.Year wrapper function}
\usage{
as.Year(international_date)
}
\arguments{
\item{international_date}{the date to convert into a year}
}
\description{
This function is a wrapper which formats a data as a Year
}
\examples{

as.Year(as.International.Date("1992/1/1"))

}
\keyword{time}
