% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Generic.R
\name{Convert_DMS_to_Decimal}
\alias{Convert_DMS_to_Decimal}
\title{Convert Degree-Minutes-Seconds coordinates to Decimal Coordinates.}
\usage{
Convert_DMS_to_Decimal(angle, split_string = " ")
}
\arguments{
\item{angle}{the angle to convert}

\item{split_string}{the characters used to separate}
}
\description{
This function converts the format of a location.
}
\examples{

Convert_DMS_to_Decimal("02131000")

}
\keyword{coordinates}
