% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaySiteInfo.R
\name{displaySiteInfo}
\alias{displaySiteInfo}
\title{displaySiteInfo}
\usage{
displaySiteInfo(siteID, HTML = FALSE)
}
\arguments{
\item{siteID}{The USGS site identified for stream flow analysis}

\item{HTML}{HTML = TRUE if output is to HTML}
}
\description{
This function displays the output of the siteInfo function in
a table using the grid.Extra function.
}
\details{
HTML - this is set to TRUE if output is to HTML
}
\examples{

displaySiteInfo("02131000")



}
\keyword{USGS}
