% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawMaps.R
\name{displaySiteLocation}
\alias{displaySiteLocation}
\title{displaySiteLocation}
\usage{
displaySiteLocation(siteID, maptype = "satellite", zoom = 14)
}
\arguments{
\item{siteID}{the USGS site identified for stream flow analysis}

\item{maptype}{the Google Maptype to use}

\item{zoom}{the zoom level on the map}
}
\description{
This function displays the output of the stationInfo function in
a table using the ggmap function.
}
\details{
This function imports the location of the site from the 
USGS database then pins it on a map from Google

credit for this function goes to this link
https://ryanpeek.github.io/2016-09-28-static_maps_in_R/ 
accessed July 2018
}
\examples{

displaySiteLocation("02131000")

}
\keyword{USGS}
