% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaySiteLocationInState.R
\name{displaySiteLocationInState}
\alias{displaySiteLocationInState}
\title{displaySiteLocationInState}
\usage{
displaySiteLocationInState(siteID)
}
\arguments{
\item{siteID}{The USGS site identified for stream flow analysis}
}
\description{
This function displays the output of the siteInfo function in
a table using the grid.Extra function.
}
\examples{

displaySiteLocationInState("02131000")

}
\keyword{USGS}
