% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{droplevelsPolygon}
\alias{droplevelsPolygon}
\title{droplevelsPolygon}
\usage{
droplevelsPolygon(polygon)
}
\arguments{
\item{polygon}{a SpatialPolygonsDataframe}
}
\value{
a SpatialPolygonsDataframe
}
\description{
This function removes unused levels from a polygon.
}
\details{
This is useful when a single watershed has been isolated. Information
concerning every other watershed is removed.
}
\examples{

droplevelsPolygon(polygon)

}
\keyword{internal}
