% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDailyValuesToCSV.R
\name{ExportDailyValuesToCSV}
\alias{ExportDailyValuesToCSV}
\title{ExportDailyValuesToCSV}
\usage{
ExportDailyValuesToCSV(dailyValues, fileName, metric = TRUE)
}
\arguments{
\item{dailyValues}{daily values output by the importDVs in the waterData package}

\item{fileName}{file name to save CSV file to}

\item{metric}{convert value to metric}
}
\description{
This function exports daily values output by the importDVs
function to a CSV file.
}
\details{
This function outputs to the working directory.  The directory can be set with 
the setwd command.

This function only supports discharge data.
}
\examples{

site = list(
   staid = "02131000",                     # site number
   code = "00060",                         # code for discharge 
   stat = "00003",                         # code for mean   
)

DailyValues <- do.call(importDVs, site)

ExportDailyValuesToCSV(DailyValues, 
                       paste0("Daily_Values_for_Site_", site$staid))

}
\keyword{USGS}
