% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportMonthlyValuesToCSV.R
\name{ExportMonthlyValuesToCSV}
\alias{ExportMonthlyValuesToCSV}
\title{ExportMonthlyValuesToCSV}
\usage{
ExportMonthlyValuesToCSV(monthlyValues, fileName, metric = TRUE)
}
\arguments{
\item{monthlyValues}{daily values output by the importDVs in the waterData package}

\item{fileName}{file name to save CSV file to}

\item{metric}{if TRUE the output is in metric units}
}
\description{
This function exports monthly values generated by the readNWISstat
function in the dataRetrieval package.
}
\details{
This function outputs to the working directory.  The directory can be set with 
the setwd command.

This function only supports discharge data.
}
\examples{

site = list(
   staid = "02131000",                     # site number
   code = "00060",                         # code for discharge 
   stat = "00003",                         # code for mean   
)

MonthlyValues <- readNWISstat(StationInfo$siteID,
                              parameterCd = StationInfo$site$code,
                              statReportType = "monthly") 

ExportDailyMonthlyToCSV(MonthlyValues, 
                       paste0("Monthly_Values_for_Site_", site$staid))

}
\keyword{USGS}
