% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Generic.R
\name{gaps_present}
\alias{gaps_present}
\title{gaps_present}
\usage{
gaps_present(x)
}
\arguments{
\item{x}{the data to check for gaps}
}
\description{
This function checks the data for gaps
}
\examples{

 gaps_present(c(1,2,3))
 gaps_present(c(1,2,NA))

}
\keyword{data}
\keyword{gaps}
