% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_NOAA_Stations.R
\name{Get_Data_from_NOAA_Station}
\alias{Get_Data_from_NOAA_Station}
\title{Get Data from NOAA Station.}
\usage{
Get_Data_from_NOAA_Station(
  station_id = NULL,
  var = NULL,
  date_min = NULL,
  date_max = NULL,
  refresh = FALSE,
  ...
)
}
\arguments{
\item{station_id}{the ID of the station to download data}

\item{var}{which elements (variables) to get from the station}

\item{refresh}{whether or not to refresh data for the station}
}
\description{
This is a wrapper for the ghcnd_search() function from the rnoaa package
}
\examples{

If NOAA stations have not been downloaded in a while be sure to refresh

}
\keyword{NOAA}
\keyword{stations}
