% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_NOAA_Stations.R
\name{Get_NOAA_Stations}
\alias{Get_NOAA_Stations}
\title{Get NOAA Stations.}
\usage{
Get_NOAA_Stations(
  localSave = TRUE,
  refresh = FALSE,
  localFileName = "stations.Rdata"
)
}
\arguments{
\item{localSave}{saves / retrieves information locally}

\item{refresh}{forces a refresh of cached station information}

\item{localFileName}{local name of file used to cache station information}
}
\description{
This function retrieves the NOAA stations.
}
\examples{

# If NOAA stations have not been downloaded in a while
 
stations <- Get_NOAA_Stations(refresh = TRUE)

 # If local saves are not desired
 
 stations <- Get_NOAA_Stations(localSave = FALSE)
 

}
\keyword{coordinates}
