% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNOAAGuages.R
\name{getNOAAGuages}
\alias{getNOAAGuages}
\title{getNOAAGuages}
\usage{
getNOAAGuages(siteID, plotmap = TRUE, zoomFactor = 10,
  limitToWatershed = TRUE, getElevations = TRUE)
}
\arguments{
\item{siteID}{the USGS site identified for stream flow analysis}

\item{plotmap}{TRUE if output plot is desired}

\item{zoomFactor}{the zoom factor of the output plot}
}
\description{
This function retrieves information on the hydraullic unit code 
given a USGS monitoring station ID.
}
\details{
Much of this code was drawn from this source.
http://cyberhelp.sesync.org/leaflet-in-R-lesson/course/archive.html
}
\examples{

getNOAAGuages("02131000")

}
\keyword{USGS}
