% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUSGS_HUC.R
\name{getUSGSHUC}
\alias{getUSGSHUC}
\title{getUSGSHUC}
\usage{
getUSGSHUC(siteID, plotmap = TRUE, zoomFactor = 10)
}
\arguments{
\item{siteID}{the USGS site identified for stream flow analysis}
}
\description{
This function retrieves information on the hydraullic unit code 
given a USGS monitoring station ID.
}
\details{
Much of this code was drawn from this source.
http://cyberhelp.sesync.org/leaflet-in-R-lesson/course/archive.html
}
\examples{

getUSGSHUC("02131000")

}
\keyword{USGS}
