% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getWatershedInfo}
\alias{getWatershedInfo}
\title{Get Watershed Information}
\usage{
getWatershedInfo()
}
\value{
Large SpatialPolygonsDataFrame
}
\description{
This function assembles all watershed information in the US
}
\details{
The watershed information is broken into 3 files based on regions, this internal function
binds these files together.
}
\examples{

WatershedInfo <- getWatershedInfo()
}
\keyword{internal}
