% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWatershedOfStation.R
\name{getWatershedOfStation}
\alias{getWatershedOfStation}
\title{Get Watershed of USGS Station}
\usage{
getWatershedOfStation(siteID, printMap = FALSE, exportMap = FALSE)
}
\arguments{
\item{siteID}{USGS station ID}

\item{printMap}{if TRUE this function prints a leaflet map and appends it to the output}

\item{exportMap}{if TRUE this function appends the leaflet map to the output without printing it}
}
\value{
a list of watershed properties
\item{name name}{description name of the watershed}
\item{name HUC}{description hydraullic unit code of watershed}
\item{name stationName}{description name of the USGS station} 
\item{name stationID}{description ID of the USGS station} 
\item{name stationLatitude}{description Latitude coordinate of USGS station}
\item{name stationLongitude}{description Longitude coordinate of USGS station}
\item{name polygon}{description polygon(SpatialPolygonsDataFrame) describing the watershed}
}
\description{
Given a USGS site this function retrieves the watershed polygon
}
\examples{

watershed <- getWatershedOfStation("02131000", printMap = TRUE)


}
