% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getWatershedRange}
\alias{getWatershedRange}
\title{Get distance range of watershed}
\usage{
getWatershedRange(watershedOfInterest)
}
\arguments{
\item{watershedOfInterest}{a SpatialPolygonsDataFrame showing a single watershed}
}
\value{
a distance in meters
}
\description{
This function gets a distance (in meters) which encompases the full range
of the watershed.  This function uses a diagonal
across the bounding box of the watershed.
}
\details{
The Haversine formula is used to calculate the distance in Longitude and
Latitude.
}
\examples{

getWatershedRange(watershedOfInterest)

}
\keyword{internal}
