% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapWatershed.R
\name{mapWatershed}
\alias{mapWatershed}
\title{mapWatershed}
\usage{
mapWatershed(
  watershedPolygon,
  USGSstationLatitude,
  USGSstationLongitude,
  NOAAstationsLatitude = NULL,
  NOAAstationsLongitude = NULL,
  USGSStationName = "",
  NOAAStationsNames = "",
  watershedMap = NULL,
  zoomFactor = 10,
  printMap = FALSE
)
}
\arguments{
\item{watershedPolygon}{a spatialPoints class defining the polygon of the watershed}

\item{USGSstationLatitude}{the latitude coordinate of the USGS station}

\item{USGSstationLongitude}{the longitude coordinate of the USGS station}

\item{NOAAstationsLatitude}{the latitude coordinates of the NOAA stations on the watershed}

\item{NOAAstationsLongitude}{the longitude coordinates of the NOAA stations on the watershed}

\item{USGSStationName}{the name of the USGS station}

\item{watershedMap}{the map of the watershed}

\item{zoomFactor}{the factor which to zoom in on a map}

\item{printMap}{set printMap to TRUE to output a map. This may not work in R markdown.}

\item{NOAAStationsName}{the name of the NOAA station}
}
\value{
a leaflet map of the watershed
}
\description{
This is a raw method of plotting a watershed.  This function is internal
as too many details need to be specified to get this function to work.
}
\keyword{internal}
