% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAAstations.R
\docType{data}
\name{NOAAstations}
\alias{NOAAstations}
\title{Location of NOAA Stations.}
\format{dataframe
\describe{
  ...
}}
\usage{
NOAAstations
}
\description{
This data is generated by the rnoaa package.
}
\details{
# NOAAstations <- ghcnd_stations() # Takes a while to run

# save(NOAAstations, "NOAAstations.rda")

This data is up to date as of July 2018
}
\keyword{datasets}
