% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitationCoverageStats.R
\name{precipitationCoverageStats}
\alias{precipitationCoverageStats}
\title{Precipitation coverage statistics}
\usage{
precipitationCoverageStats(precipitationData, type = "html")
}
\arguments{
\item{precipitationData}{a dataframe output by meteo_pull_monitors 
from the rnoaa package}
}
\value{
print output of a table with coverage statistics
}
\description{
Precipitation data is known to have gaps for any number of reasons
 
 This function is intended to produce a table of outputs concerning
 how well the range of a particular statistic is covered.
}
\examples{

precipitationCoverageStats(precipitationData)

}
