% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Generic.R
\name{remove_attributes}
\alias{remove_attributes}
\title{remove_attributes}
\usage{
remove_attributes(x)
}
\arguments{
\item{x}{the variable which to remove the attributes}
}
\description{
This function clears attributes
 
 The parameters can be shorted to lat, lon if space is short.
}
\examples{

df <- list(
  fun = function(x){sum(x)},
  fun1 = function(x){mean(x)},
  fun2 = function(x){sd(x)} 
)

df$fun <- remove_attributes(df$fun)

df <- lapply(df, remove_attributes)


}
\keyword{attr}
\keyword{attributes}
