% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_NOAA_Stations.R
\name{Select_NOAA_Stations}
\alias{Select_NOAA_Stations}
\title{Select NOAA Stations.}
\usage{
Select_NOAA_Stations(
  localSave = FALSE,
  refresh = FALSE,
  localFileName = "stations.Rdata",
  selectElements = NULL,
  beginYear = NULL,
  endYear = NULL,
  unique_stations = FALSE,
  unique_cols = NULL,
  NOAA_variables = c("PRCP")
)
}
\arguments{
\item{localSave}{saves / retrieves information locally}

\item{refresh}{forces a refresh of cached station information}

\item{localFileName}{local name of file used to cache station information}

\item{selectElements}{list of elements to limit select}

\item{beginYear}{the earliest year of data collection}

\item{endYear}{the last year of data collection}

\item{unique_stations}{reorganize so output only includes unique stations}

\item{unique_cols}{when collapsing to only unique stations, this is the 
list of which rows to include.}

\item{NOAA_variables}{collapse stations to only unique stations in list}
}
\description{
This function retrieves the NOAA stations, like Get_NOAA_Stations,
 then selects a group of stations based on certain properties.
}
\examples{

# If NOAA stations have not been downloaded in a while
 
stations <- Select_NOAA_Stations(refresh = TRUE)

 # If local saves are not desired
 
 stations <- Select_NOAA_Stations(localSave = FALSE)
 
 
 

}
\keyword{coordinates}
