% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{spatialPointsLatLng}
\alias{spatialPointsLatLng}
\title{spatialPointsLatLng}
\usage{
spatialPointsLatLng(latitude, longitude, proj4stringAttributes = NULL)
}
\arguments{
\item{latitude}{latitude coordinates}

\item{longitude}{longitude coordinates}

\item{proj4stringAttributes}{(optional) which projection to copy attributes from}
}
\value{
a SpatialPolygonsDataframe
}
\description{
This function is a wrapper for the spatialPoints function in the sp package
}
\details{
This turns lattitude and longitude coordinates into a spaital projection
}
\examples{

spatialPointsLatLng (latitude, longitude)
}
