% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterData_URL.R
\docType{data}
\name{waterData_URL}
\alias{waterData_URL}
\title{URLs associated with the waterData Support Package.}
\format{a list of URLS
\describe{
  \item{name National_Hydrography_Dataset}{URL of national hydrography dataset}
  ...
}}
\usage{
waterData_URL
}
\description{
URLs associated with the waterData Support Package.
}
\keyword{datasets}
