% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershedInfo.R
\docType{data}
\name{watershedInfo}
\alias{watershedInfo}
\title{Hydraullic Unit Codes from Watershed Boundary Dataset.}
\format{A Large SpatialPolygonsDataFrame with 2158 elements:
\describe{
  \item{AREA}{}
  \item{PERIMETER}{}
  \item{HUC250K_}{}
  \item{HUC250K_ID}{}
  \item{HUC_CODE}{}
  \item{HUC_NAME}{}
  \item{REG}{}
  \item{SUB}{}
  \item{ACC}{}
  \item{CAT}{}
  ...
}}
\source{
\url{https://nhd.usgs.gov/wbd.html}
}
\usage{
watershedInfo
}
\description{
This is an imported shapefile containing the watershed boundaries in the US.
}
\details{
To generate this dataset download the huc250k dataset from the
USGS website

# library(sp)
# library(rgdal)

## Set working directory to unzipped shape file directory

# watershedInfo <- spTransform(readOGR(".", "huc250k"), 
#                              CRS("+proj=longlat +datum=WGS84"))

wIgroup1 <- watershedInfo[as.numeric(watershedInfo$REG) <= 6,]
wIgroup2 <- watershedInfo[as.numeric(watershedInfo$REG) >= 7 &
                            as.numeric(watershedInfo$REG) <= 12 ,]
wIgroup3 <- watershedInfo[as.numeric(watershedInfo$REG) >= 13,]

# save(wIgroup1, file="wIgroup1.RDA")
# save(wIgroup2, file="wIgroup2.RDA")
# save(wIgroup3, file="wIgroup3.RDA")

wateshedInfo can be called as followed
watershedInfo <- rbind(wIgroup1, wIgroup2, wIgroup3)
}
\keyword{datasets}
