% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNOAAGuages.R
\name{getNOAAGuages}
\alias{getNOAAGuages}
\title{getNOAAGuages}
\usage{
getNOAAGuages(siteID, plotmap = TRUE, zoomFactor = 10,
  limitToWatershed = TRUE, getElevations = TRUE, leafletmap = NULL)
}
\arguments{
\item{siteID}{the USGS site identified for stream flow analysis}

\item{plotmap}{TRUE if output plot is desired}

\item{zoomFactor}{the zoom factor of the output plot}

\item{limitToWatershed}{only pull datapoints on the watershed}

\item{getElevations}{retrieve elevations of datapoints}

\item{leafletmap}{pass a leaflet map arguements, note the output changes to a 
list if this parameter is invoked}
}
\description{
This function retrieves information on the hydraullic unit code 
given a USGS monitoring station ID.
}
\details{
Much of this code was drawn from this source.
http://cyberhelp.sesync.org/leaflet-in-R-lesson/course/archive.html
}
\examples{

getNOAAGuages("02131000")

}
\keyword{NOAA}
\keyword{USGS}
\keyword{mapping}
\keyword{precipitation}
